Add step-by-step ACO visualization with raylib

+++ Graph.cpp
@@ -1,20 +1,16 @@
 #include "Graph.hpp"
 
-Graph::Graph(int size) 
-	: peso(size, std::vector<double>(size, 0.0))
+Graph::Graph(int size)
+    : peso(size, std::vector<double>(size, 0.0))
 {
-
 }
 
-int Graph::size()const
+int Graph::size() const
 {
-	return peso.size();
+    return static_cast<int>(peso.size());
 }
 
 double Graph::getPeso(int origem, int destino) const
 {
-	return peso[origem][destino];
+    return peso[origem][destino];
 }
-
-
-
